\name{mg_nodes}
\alias{mg_nodes}
\alias{mg_edges}
\alias{mg_metadata}
\alias{mg_sources}
\title{Access mechgraph tables and metadata}
\usage{
mg_nodes(x)

mg_edges(x)

mg_metadata(x)

mg_sources(x)
}
\arguments{
\item{x}{A mechgraph object.}
}
\description{
Return the node table, edge table, metadata list, or unique source labels from a mechgraph object.
}
